% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIPSr-package.R
\docType{package}
\name{AIPSr-package}
\alias{AIPSr}
\alias{AIPSr-package}
\title{AIPSr: AI-photoswitchable screening (AI-PS)}
\description{
This project permits the development and utilization of a pooled optical screen by making use of photoactivation and sorting. AI-PS algorithms provide a structure for segmenting cells, training machine learning algorithms to categorize cellular phenotypes, and applying these for executing a screen on the fly. To learn more, take a look at Kanfer et al., 2021.
}
\keyword{internal}
