% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM_model.R
\name{extractFeatures}
\alias{extractFeatures}
\title{2.1 Measures cell features}
\usage{
extractFeatures(x, mask_cyto, label_class = "Positive")
}
\arguments{
\item{x}{(Normalized) image of cytosolic/phenotype channel. Value returned from segmentCyto (out$norm) or user generated via EBImage.}

\item{mask_cyto}{Cell segmentation mask output from segmentCyto}

\item{label_class}{String. Positive or Negative.}
}
\value{
List of table, and row-names table (?) with features (cytosolic img) for each segmented cell
}
\description{
2.1 Measures cell features
}
