% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM_model.R
\name{featuresPCA}
\alias{featuresPCA}
\title{2.4 PCA Analysis: features reduction}
\usage{
featuresPCA(
  Table_class_train,
  cartesian_lim_x = c(-12, 10),
  cartesian_lim_y = c(-10, 5),
  font_size = 14
)
}
\description{
2.4 PCA Analysis: features reduction
}
