% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNN_model.R
\name{genImageSet}
\alias{genImageSet}
\title{3.1 Generation and upload of single cell images for CNN training}
\usage{
genImageSet(cseg, base_dir, pheno0, pheno1, int)
}
\arguments{
\item{pheno0}{String. Name of control (normal) phenotype}

\item{pheno1}{String. Name of positive phenotype}

\item{label_class}{String. Classification of negative or positive phenotype?}
}
\description{
3.1 Generation and upload of single cell images for CNN training
}
