% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{LoadImage}
\alias{LoadImage}
\title{Loads Image
This function loads an image from file-name, then normalizes each channel}
\usage{
LoadImage(file_name)
}
\arguments{
\item{file_name}{path to image file (string)}
}
\value{
Image object (EBImage)
}
\description{
Loads Image
This function loads an image from file-name, then normalizes each channel
}
