% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_segmentation.R
\name{mask_ch}
\alias{mask_ch}
\title{Make pre-segmented mask for single-channel image}
\usage{
mask_ch(img, filter_size, offset, opensize, shape)
}
\arguments{
\item{img}{processed or normalized single-channel image}

\item{filter_size}{size (x and y) of threshold window}

\item{offset}{?}

\item{opensize}{?}

\item{shape}{shape of brush used to open mask (string)}
}
\description{
Make pre-segmented mask for single-channel image
}
