% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM_model.R
\name{modelSVM}
\alias{modelSVM}
\title{2.3 SVM model}
\usage{
modelSVM(total_train, kernel_linear = TRUE, cost = 10, degree = 45)
}
\arguments{
\item{total_train}{Combined table of both positive and negative classified cells}

\item{kernel_linear}{??}

\item{cost}{?}

\item{degree}{?}
}
\value{
model accuracy **and model??
}
\description{
2.3 SVM model
}
