% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_segmentation.R
\name{norm_ch}
\alias{norm_ch}
\title{Return single channel from loaded image}
\usage{
norm_ch(image, index, minmaxnorm = TRUE)
}
\value{
single channel image
}
\description{
Return single channel from loaded image
}
