% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM_model.R
\name{pickCells}
\alias{pickCells}
\title{2.2 Choose cells for classification}
\usage{
pickCells(
  mask_nuc,
  x,
  xy_nuc,
  Ts.mix,
  int,
  font_size = 0.7,
  label_class = "Postive",
  display_select = TRUE
)
}
\description{
2.2 Choose cells for classification
}
