% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_segmentation.R
\name{segmentCyto}
\alias{segmentCyto}
\title{1.2 Cell Segmentation
Segments cells using nuclei seeds and cytosolic marker}
\usage{
segmentCyto(
  x,
  y,
  index = 2,
  int = 40,
  filter_size = 10,
  offset = 0.1,
  size_smooth = 19,
  opensize = 7,
  largeobj = 30000,
  minmaxnorm = TRUE
)
}
\arguments{
\item{x}{loaded image}

\item{y}{nseg returned from segmentNucleus}

\item{index}{Integer. Which channel (1,2...) has cytosolic / phenotype image}

\item{largeobj}{Integer. Maximum size (area) of cell included in final segmentation.}
}
\value{
List. Nucleus segmentation mask (nuc_seg), cytosolic segmentation mask (seg), mask features for nucleus segmentation (xy_nuc) and cytosolic segmentation (xy_cseg), normalized single cytosolic channel image (norm)
}
\description{
1.2 Cell Segmentation
Segments cells using nuclei seeds and cytosolic marker
}
