% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_segmentation.R
\name{segmentNucleus}
\alias{segmentNucleus}
\title{1.1 Nucleus Segmentation}
\usage{
segmentNucleus(
  image,
  index = 1,
  minmaxnorm = TRUE,
  int = 2,
  filter_size = 16,
  offset = 0.04,
  opensize = 3,
  small_obj = 30,
  use_watershed = FALSE,
  distmap_value = 2,
  rm_outliers = TRUE,
  out_p = 0.95,
  displaymasks = TRUE
)
}
\arguments{
\item{image}{EBImage object}

\item{index}{select which channel (1=ch1, 2=ch2..) integer value}

\item{minmaxnorm}{whether or not to normalize each channel by min/max intensity (TRUE or FALSE)}

\item{int}{intensity of image (integer)}

\item{filter_size}{size (x and y) of threshold window}

\item{offset}{?}

\item{opensize}{?}

\item{small_obj}{Integer, all objects smaller than value are removed}

\item{use_watershed}{TRUE/FALSE, whether to use EBImage watershed for segmenting, instead of simpler bwlabel function}

\item{distmap_value}{If watershed=TRUE, edge sensitivity of segmentation}

\item{rm_outliers}{TRUE/FALSE ?}

\item{out_p}{?}
}
\value{
list: segmented nuclei mask, xy features, (display masks on image)
}
\description{
1.1 Nucleus Segmentation
}
