% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNN_model.R
\name{splitDirs}
\alias{splitDirs}
\title{3.2 Create testing, training directories (split based off of user-inputed ratio) and move files to appropriate destinations
Create files into testing and training sets, create test directory. Remaining files in original base_dir=train_dir}
\usage{
splitDirs(base_dir, pheno0, pheno1, test_ratio)
}
\arguments{
\item{base_dir}{}

\item{pheno0}{String. Name of control (normal) phenotype}

\item{pheno1}{String. Name of positive phenotype}

\item{test_ratio}{Number between 0-1. Proportion of images to relocate to testing directory}
}
\description{
3.2 Create testing, training directories (split based off of user-inputed ratio) and move files to appropriate destinations
Create files into testing and training sets, create test directory. Remaining files in original base_dir=train_dir
}
